+++ main.go
@@ -1,23 +1,17 @@
 package main
 
 import (
+	"os"
+
+	"github.com/OdairPianta/gryphon_api/docs"
 	"github.com/OdairPianta/gryphon_api/models"
 	"github.com/OdairPianta/gryphon_api/routes"
 )
 
 // swagger embed files
-
-// @title           Gryphon API
-// @version         0.1.1
-// @description     Gryphon API is a REST API that allows send files and manage files.
-
 // @contact.name   API Support
 // @contact.url    https://spotec.app/contato/
 // @contact.email  contact@spotec.app
-
-// @host      localhost:8083
-// @BasePath  /api
-
 // @securityDefinitions.apikey Bearer
 // @in header
 // @name Authorization
@@ -26,6 +20,14 @@ import (
 // @externalDocs.description  OpenAPI
 // @externalDocs.url          https://swagger.io/resources/open-api/
 func main() {
-	models.ConnectDatabase()
+	models.InitApp()
+
+	docs.SwaggerInfo.Title = "Gryphon API"
+	docs.SwaggerInfo.Description = "Gryphon API is a REST API that allows send files and manage files."
+	docs.SwaggerInfo.Version = "1.1.2"
+	docs.SwaggerInfo.Host = os.Getenv("APP_HOST") + ":" + os.Getenv("APP_PORT")
+	docs.SwaggerInfo.BasePath = "/api"
+	docs.SwaggerInfo.Schemes = []string{"http"}
+
 	routes.HandleRequest()
 }