+++ http/requests/create_base64_file_input.go
@@ -1,0 +1,7 @@
+package requests
+
+type CreateBase64FileInput struct {
+	Base64    string `json:"base64" binding:"required"`
+	Extension string `json:"extension" binding:"required"`
+	FilePath  string `json:"file_path"`
+}